+++ Assets/Prefabs/ChimeTree.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 37dc73594dc9c3d41872a5519d69c5f7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
