+++ Assets/QuickOutline.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 395aa5fc9a8f5d5419e880a14cd773b8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
