+++ Assets/QuickOutline/Readme.txt.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 16a82a9714ebab5479955d20c1dba569
+timeCreated: 1522619008
+licenseType: Store
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
