+++ Assets/QuickOutline/Resources.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 70fd40674751a8042a8b9b2e8d9f915f
+folderAsset: yes
+timeCreated: 1522559128
+licenseType: Store
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
