+++ Assets/QuickOutline/Resources/Materials.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 80ac8e52d3c31a94babd161e86bc6b97
+folderAsset: yes
+timeCreated: 1522559139
+licenseType: Store
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
