+++ Assets/QuickOutline/Resources/Materials/OutlineFill.mat.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 311313efa011949e98b6761d652ad13c
+timeCreated: 1520576285
+licenseType: Store
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
