+++ Assets/QuickOutline/Resources/Shaders.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6a63caa2b0e993043a42c11f35ff2d1a
+folderAsset: yes
+timeCreated: 1522559134
+licenseType: Store
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
