+++ Assets/QuickOutline/Resources/Shaders/OutlineFill.shader.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 4e76d4023d7e0411297c670f878973e2
+timeCreated: 1520575782
+licenseType: Store
+ShaderImporter:
+  externalObjects: {}
+  defaultTextures: []
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
