+++ Assets/QuickOutline/Scripts.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 3ddbd65d69a9f0b48bab4fe96a1fe099
+folderAsset: yes
+timeCreated: 1522559122
+licenseType: Store
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
