--- Assets/Scripts/Outline.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 5fea29bb7c508c244a1f805a5fd3fc4d
-timeCreated: 1522369084
-licenseType: Store
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
