--- Assets/Scripts/Testbug.cs
@@ -1,10 +1,0 @@
-
-using UnityEngine;
-
-public class Testbug : MonoBehaviour
-{
-    void Update()
-    {
-        Debug.Log("AAAAAAAAAAAAAAAA");
-    }
-}
