--- Assets/Scripts/Testbug.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fcbc62652ddc92a498fccd95a43087dc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
