--- Assets/Scripts/player/QuickOutlineReadme.txt.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5933bfd39d7a5b843a0ed821f85bca19
-timeCreated: 1522619008
-licenseType: Store
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
