+++ Assets/Scripts/player/playerLookAtHighlighter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0ff712599ddc78645a9aab79300bdff0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
