--- Assets/Shaders/OutlineMask.shader
@@ -1,33 +1,0 @@
-﻿//
-//  OutlineMask.shader
-//  QuickOutline
-//
-//  Created by Chris Nolet on 2/21/18.
-//  Copyright © 2018 Chris Nolet. All rights reserved.
-//
-
-Shader "Custom/Outline Mask" {
-  Properties {
-    [Enum(UnityEngine.Rendering.CompareFunction)] _ZTest("ZTest", Float) = 0
-  }
-
-  SubShader {
-    Tags {
-      "Queue" = "Transparent+100"
-      "RenderType" = "Transparent"
-    }
-
-    Pass {
-      Name "Mask"
-      Cull Off
-      ZTest [_ZTest]
-      ZWrite Off
-      ColorMask 0
-
-      Stencil {
-        Ref 1
-        Pass Replace
-      }
-    }
-  }
-}
