--- Assets/Shaders/OutlineMask.shader.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 341b058cd7dee4f5cba5cc59a513619e
-timeCreated: 1520575773
-licenseType: Store
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
